requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-docx>=1.1.0
openpyxl>=3.1.0
openai>=1.60.0
